F [#PS-45] Split disputes into classes, add collection and Dispute.collection_from_payment_service
