+++ lib/payment/all.rb
@@ -1,0 +1,6 @@
+require_relative 'connector'
+require_relative 'organization_connector'
+require_relative 'response_classes/stripe_hash'
+require_relative 'response_classes/collection'
+require_relative 'response_classes/merchant_stripe'
+require_relative 'response_classes/dispute'
