--- lib/payment/all_connectors.rb
@@ -1,2 +1,0 @@
-require_relative 'connector'
-require_relative 'organization_connector'
