+++ lib/payment/version.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 module ShorePayment # :nodoc:
-  VERSION = '0.2.4'
+  VERSION = '0.2.5'
 end
