+++ spec/payment/merchant_stripe_spec.rb
@@ -437,54 +437,4 @@ describe ShorePayment::MerchantPayment do
       expect(subject.charges.first < subject.charges.last).to eq(true)
     end
   end
-
-  context '#disputes' do
-    let(:dispute) do
-      ShorePayment::Dispute.new(
-        payment_service_dispute_response(oid, {})['dispute']
-      )
-    end
-
-    describe 'attributes' do
-      it { expect(dispute).to respond_to(:id) }
-      it { expect(dispute).to respond_to(:status) }
-      it { expect(dispute).to respond_to(:reason) }
-      it { expect(dispute).to respond_to(:amount_cents) }
-      it { expect(dispute).to respond_to(:currency) }
-      it { expect(dispute).to respond_to(:created_at) }
-      it { expect(dispute).to respond_to(:organization_id) }
-      it { expect(dispute).to respond_to(:due_by) }
-      it { expect(dispute).to respond_to(:has_evidence) }
-      it { expect(dispute).to respond_to(:past_due) }
-      it { expect(dispute).to respond_to(:submission_count) }
-      it { expect(dispute).to respond_to(:evidence) }
-    end
-
-    describe 'methods' do
-      it { expect(dispute).to respond_to(:update) }
-    end
-
-    context '#evidence' do
-      let(:evidence) { dispute.evidence }
-
-      describe 'attributes' do
-        it { expect(evidence).to respond_to(:product_description) }
-        it { expect(evidence).to respond_to(:customer_name) }
-        it { expect(evidence).to respond_to(:customer_email_address) }
-        it { expect(evidence).to respond_to(:billing_address) }
-        it { expect(evidence).to respond_to(:receipt) }
-        it { expect(evidence).to respond_to(:customer_signature) }
-        it { expect(evidence).to respond_to(:customer_communication) }
-        it { expect(evidence).to respond_to(:uncategorized_file) }
-        it { expect(evidence).to respond_to(:uncategorized_text) }
-        it { expect(evidence).to respond_to(:service_date) }
-        it { expect(evidence).to respond_to(:service_documentation) }
-        it { expect(evidence).to respond_to(:shipping_address) }
-        it { expect(evidence).to respond_to(:shipping_carrier) }
-        it { expect(evidence).to respond_to(:shipping_date) }
-        it { expect(evidence).to respond_to(:shipping_documentation) }
-        it { expect(evidence).to respond_to(:shipping_tracking_number) }
-      end
-    end
-  end
 end
